+++ packages/breadboard/src/data/types.ts
@@ -5,6 +5,7 @@
  */
 
 import { HarnessRunResult } from "../harness/types.js";
+import { ReanimationState } from "../run/types.js";
 
 export type FunctionCallCapabilityPart = {
   functionCall: {
@@ -71,6 +72,10 @@ export type SerializedDataStoreGroup = SerializedStoredData[];
 /**
  * A provider that handles storing and retrieving data.
  */
+
+export type RunURL = string;
+export type RunTimestamp = number;
+
 export type DataStoreProvider = {
   store(data: InlineDataCapabilityPart): Promise<StoredData>;
   retrieve(handle: DataStoreHandle): Promise<StoredData>;
@@ -79,12 +84,17 @@ export type DataStoreProvider = {
 };
 
 export type RunStore = {
-  start(storeId: string, limit?: number): Promise<string>;
-  write(result: HarnessRunResult): Promise<void>;
-  stop(): Promise<void>;
-  abort(): Promise<void>;
-  drop(): Promise<void>;
-  getNewestRuns(limit: number): Promise<HarnessRunResult[][]>;
+  start(url: RunURL): Promise<RunTimestamp>;
+  write(
+    url: RunURL,
+    timestamp: RunTimestamp,
+    result: HarnessRunResult
+  ): Promise<void>;
+  stop(url: RunURL, timestamp: RunTimestamp): Promise<void>;
+  abort(url: RunURL, timestamp: RunTimestamp): Promise<void>;
+  drop(url?: RunURL): Promise<void>;
+  truncate(url: RunURL, limit: number): Promise<void>;
+  getStoredRuns(url: RunURL): Promise<Map<RunTimestamp, HarnessRunResult[]>>;
 };
 
 export type DataStore = {
@@ -100,4 +110,9 @@ export type DataStore = {
     storeId?: string
   ): Promise<SerializedDataStoreGroup | null>;
   store(blob: Blob, storeId?: string): Promise<StoredDataCapabilityPart>;
+};
+
+export type StateStore = {
+  load(key?: string): Promise<ReanimationState | undefined>;
+  save(state: ReanimationState): Promise<string>;
 };