+++ packages/breadboard/tests/new/grammar/serialize-and-run-with-old-runner.ts
@@ -11,8 +11,8 @@ import { Serializeable } from "../../../src/new/runner/types.js";
 import {
   InputValues,
   OutputValues,
-  BoardRunner,
   asRuntimeKit,
+  invokeGraph,
 } from "../../../src/index.js";
 
 import { TestKit, testKit } from "../../helpers/_test-kit.js";
@@ -21,8 +21,8 @@ async function serializeAndRunGraph(
   graph: Serializeable,
   inputs: InputValues
 ): Promise<OutputValues> {
-  const board = await BoardRunner.fromGraphDescriptor(await graph.serialize());
-  return board.runOnce(inputs, { kits: [asRuntimeKit(TestKit)] });
+  const board = await graph.serialize();
+  return invokeGraph(board, inputs, { kits: [asRuntimeKit(TestKit)] });
 }
 
 test("simplest graph", async (t) => {